Move to Ledger.Extra. Add mustPayToAddress
